End special treatment for AppDynamics
As of AppDynamics Agent 4.2.4 there is no need to have
a special node version which supports AppDynamics.

See the following post:

https://docs.appdynamics.com/display/PRO42/Node.js+Supported+Environments

As of 4.2.4, the Node.js Agent supports all major and minor versions of
Node.js v0, v4, v5, and v6. Further, the agent no longer needs to be
updated for each new minor version update of Node.js.